contract-verifier: eagerify handler facts, dispose trees per file
The end-to-end fix for WASM heap exhaustion on large monorepos. Verified
on payloadcms/payload (4493 TS files) end-to-end: 69s wall-clock, exit 0,
zero Aborted(), 3 drifts matching the prior run. Replaces PR #532 which
deferred-disposal-only and still crashed at payload scale.

## What broke

ExtractedOperation carried `handlerBody?: SyntaxNode` (a tree-sitter node
into the source tree) plus `handlerSource?: string`. Two comparison-time
consumers walked it:

  - effect/emission-facts.ts extractEmissionFacts(ops) — for
    compareEffectGroup. Did the full emit() / dynamic-emit /
    failure-block / branch-emit AST walk per operation.
  - comparator/authorization-rule.ts handlerHasOwnershipCheck(body,
    source, fieldName) — for compareAuthorizationRule. Walked the body
    looking for equality comparisons with req.auth.* / current_user.*.

Because the tree had to stay alive until comparators ran, no tree could
be freed per-file. PR #531 tried per-file delete() anyway and crashed
with `RuntimeError: memory access out of bounds`. PR #532 deferred all
disposal to end-of-verify(), keeping every tree in WASM heap until then,
which still exhausts the heap on payload-scale repos (35+ min then
Aborted() flood).

## What this does

New module extractor/handler-facts.ts owns the eager extraction. At the
moment we have the handler body's AST, we compute and store as plain data:

  emission: OperationEmission             // Sets/Maps of strings + line numbers
  ownershipCheckCandidates:
    Array<{ resourceField: string; line: number }>

These replace handlerBody/handlerSource on ExtractedOperation. The three
producers (operation.ts, operation-fastapi.ts, file-based-routes.ts) now
call extractHandlerFacts() at handler-resolution time.

Consumers rewired:

  - emission-facts.ts is now a 50-line re-keyer over op.emission. The
    250 lines of AST-walking logic moved verbatim into handler-facts.ts
    (same semantics; preserved tests).
  - authorization-rule.ts no longer walks AST. Comparator body becomes:
        op.ownershipCheckCandidates?.some(c => c.resourceField === fieldName)
    handlerHasOwnershipCheck / matchesAuthSide / matchesResourceField
    deleted from comparator/ (their auth-side / terminal-property logic
    lives in handler-facts.ts now).

With nothing downstream holding tree nodes, per-file dispose is safe.
All five walkers updated:

  - source-walker.ts eachParsedSource: dispose in finally after visit().
    Removed trackTree / disposeAllTrackedTrees registry (PR #532's
    safety net is no longer needed).
  - extractor/index.ts extractOperationsFromDir: dispose in finally.
  - file-based-routes.ts: dispose in finally.
  - auth-presence.ts: dispose in finally (UseEdge was already plain).
  - idempotency-presence.ts: refactored two-pass model. Was Pass 1
    parsed everything + stored every tree in fileTrees map until Pass 2
    walked routes — that map alone held 4500 trees on payload, the same
    failure shape. Now Pass 1 collects BOTH bindings AND
    PendingRoute { declarationLine, middlewareIdents,
    handlerReadsHeader: bool } per file, disposes the tree, and Pass 2
    runs over plain data (cross-file aware-function lookup).

Removed the disposeAllTrackedTrees try/finally wrappers from verify()
and infer() — no longer needed with per-file disposal.

## Validation

  - tests/contract-verifier/  389/389 pass. Most importantly:
    verify-end-to-end.test.ts exercises both compareAuthorizationRule
    (the IL-DRIFT marker `AuthorizationRule:order.owner-only / GET
    /api/orders/{id} / missing-ownership-check` lives in the fixture)
    and the EffectGroup path. Drift set unchanged.
  - End-to-end on payloadcms/payload at the pinned ref
    `a6e494ed6b8488e186e7e1b02d3bb0a9309c0bbd` with the contracts
    fetched from `origin/claude/drift-fp-store/payloadcms-payload`:
        wall-clock     69s   (was crashing after 35+ min)
        exit code      0
        Aborted() msgs 0
        drifts         3   (all named-constant/no-code-counterpart [info])
        fp             0
    The drift set matches PR #531's reported numbers byte-for-byte.

## Diff size

11 files changed, 208 insertions(+), 524 deletions(-). Net negative
because the AST-walking logic in emission-facts.ts collapsed once it
moved to handler-facts.ts and the comparator no longer needed its own
copy of the auth-side detector.

External impact: zero. ExtractedOperation / handlerBody are only used
inside packages/contract-verifier/src/. Grepped the whole monorepo to
confirm.
